feat: 홈 라우터 삭제
