+++ src/Components/Navigation.js
@@ -4,7 +4,6 @@ import { Link, withRouter } from 'react-router-dom';
 const Navigation = withRouter(({ location: { pathname } }) => {
   return (
     <>
-      <Link to="/">Home</Link>
       <Link to="/product">Product</Link>
       <Link to="/recentList">RecentList</Link>
     </>
