+++ src/Components/Router.js
@@ -2,7 +2,6 @@ import React, { Component } from 'react';
 import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom';
 //
 import Navigation from './Navigation';
-import Home from '../Routers/Home';
 import Product from '../Routers/Product';
 import Detail from '../Routers/Detail';
 import Recent from '../Routers/Recent';
@@ -13,7 +12,6 @@ class Router extends Component {
       <BrowserRouter>
         <Navigation />
         <Switch>
-          <Route path="/" exact component={Home}></Route>
           <Route path="/product" exact component={Product}></Route>
           <Route path="/product/:id" component={Detail}></Route>
           <Route path="/recentList" component={Recent}></Route>
