+++ src/Routers/Detail/DetailPresenter.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-const HomePresenter = (props) => {
+const DetailPresenter = (props) => {
   return <div>상세화면입니다</div>;
 };
 
-export default HomePresenter;
+export default DetailPresenter;
