--- src/Routers/Home/index.js
@@ -1,3 +1,0 @@
-import HomeContainer from './HomeContainer';
-
-export default HomeContainer;
