+++ src/Routers/Product/ProductPresenter.js
@@ -1,7 +1,20 @@
 import React from 'react';
 
-const ProductPresenter = (props) => {
-  return <div>제품화면입니다</div>;
+const ProductPresenter = ({ data }) => {
+  return (
+    <ul>
+      {data.map((item) => (
+        <div key={item.id} style={{ margin: 30 }}>
+          <li>{item.id}</li>
+          <li>{item.title}</li>
+          <li>{item.brand}</li>
+          <li>{item.price}</li>
+          <li>{item.date}</li>
+          <li>{String(item.liked)}</li>
+        </div>
+      ))}
+    </ul>
+  );
 };
 
 export default ProductPresenter;
